CM-36351 - Add "Ignore this violation" for secrets
