+++ src/extension/Cycode.VisualStudio.Extension.Shared/Cli/DTO/CliIgnoreType.cs
@@ -1,0 +1,7 @@
+﻿namespace Cycode.VisualStudio.Extension.Shared.Cli.DTO;
+
+public enum CliIgnoreType {
+    Value,
+    Path,
+    Rule,
+}
