+++ src/extension/Cycode.VisualStudio.Extension.Shared/Components/ViolationCards/SecretViolationCardControl.xaml
@@ -17,6 +17,8 @@
                 <RowDefinition />
                 <RowDefinition />
                 <RowDefinition />
+                <RowDefinition />
+                <RowDefinition />
             </Grid.RowDefinitions>
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="2*" />
@@ -47,6 +49,14 @@
 
             <t:Header Grid.Row="10" Grid.Column="0" Grid.ColumnSpan="2" Text="Company Guidelines" />
             <t:Markdown x:Name="CompanyGuidelines" Grid.Row="11" Grid.Column="0" Grid.ColumnSpan="2" />
+            
+            <t:Hr Grid.Row="12" Grid.Column="0" Grid.ColumnSpan="2" />
+            <Button Grid.Row="13" Grid.Column="1" 
+                    Content="Ignore this violation" 
+                    HorizontalAlignment="Right" 
+                    VerticalAlignment="Bottom" 
+                    Margin="0 0 10 10"
+                    Click="IgnoreButton_OnClickAsync" />
         </Grid>
     </ScrollViewer>
 </UserControl>
