+++ src/extension/Cycode.VisualStudio.Extension.Shared/Services/DocTableEventsHandlerService.cs
@@ -9,6 +9,11 @@ using Microsoft.VisualStudio.Shell.Interop;
 
 namespace Cycode.VisualStudio.Extension.Shared.Services;
 
+public interface IDocTableEventsHandlerService {
+    void Initialize(IServiceProvider serviceProvider);
+    void Deinitialize();
+}
+
 public class DocTableEventsHandlerService(
     ILoggerService logger,
     ICycodeService cycode,
