--- src/extension/Cycode.VisualStudio.Extension.Shared/Services/ErrorList/IErrorTaskCreatorService.cs
@@ -1,6 +1,0 @@
-﻿namespace Cycode.VisualStudio.Extension.Shared.Services.ErrorList;
-
-public interface IErrorTaskCreatorService {
-    Task RecreateAsync();
-    Task ClearErrorsAsync();
-}
