--- src/extension/Cycode.VisualStudio.Extension.Shared/Services/IDocTableEventsHandlerService.cs
@@ -1,6 +1,0 @@
-﻿namespace Cycode.VisualStudio.Extension.Shared.Services;
-
-public interface IDocTableEventsHandlerService {
-    void Initialize(IServiceProvider serviceProvider);
-    void Deinitialize();
-}
