--- src/extension/Cycode.VisualStudio.Extension.Shared/Services/IStateService.cs
@@ -1,8 +1,0 @@
-using Cycode.VisualStudio.Extension.Shared.DTO;
-
-namespace Cycode.VisualStudio.Extension.Shared.Services;
-
-public interface IStateService {
-    ExtensionState Load();
-    void Save();
-}
